requirements.txt
Django>=1.11
dj-database-url==0.2.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
coreapi==2.3.0
whitenoise==3.2.2
psycopg2==2.7.3.2
gunicorn==19.6.0
uWSGI==2.0.15
xmlschema==0.9.9
openpyxl==2.4.8

# for development and testing. Need to move this into seperate file
autopep8==1.3.3
coveralls==1.1
coverage==4.2
tox==2.7.0

